+++ eunbi/room-measure/backend-cloud/Dockerfile
@@ -1,5 +1,5 @@
-# Python 3.11 slim 이미지 사용
-FROM python:3.11-slim
+# Python 3.12 slim 이미지 사용
+FROM python:3.12-slim
 
 # 작업 디렉토리 설정
 WORKDIR /app
@@ -10,13 +10,13 @@ RUN apt-get update && apt-get install -y \
     && rm -rf /var/lib/apt/lists/*
 
 # Python 의존성 파일 복사
-COPY requirements.txt .
+COPY pyproject.toml .
 
 # uv 설치 (빠른 패키지 관리자)
 RUN pip install uv
 
 # 의존성 설치
-RUN uv pip install --system -r requirements.txt
+RUN uv pip install --system .
 
 # 애플리케이션 코드 복사
 COPY . .
@@ -32,5 +32,5 @@ USER appuser
 # 포트 노출
 EXPOSE 3000
 
-# 애플리케이션 실행
+# 애플리케이션 실행 (모듈화된 구조)
 CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "3000"]
