+++ eunbi/room-measure/backend-cloud/src/auth/service.py
@@ -1,0 +1,41 @@
+import jwt
+import bcrypt
+from datetime import datetime, timedelta
+from fastapi import HTTPException, Depends
+from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
+from config.settings import JWT_SECRET, JWT_ALGORITHM, JWT_EXPIRATION_HOURS
+
+security = HTTPBearer()
+
+
+def hash_password(password: str) -> str:
+    """비밀번호 해싱"""
+    return bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')
+
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    """비밀번호 검증"""
+    return bcrypt.checkpw(plain_password.encode('utf-8'), hashed_password.encode('utf-8'))
+
+
+def create_access_token(data: dict) -> str:
+    """JWT 토큰 생성"""
+    to_encode = data.copy()
+    expire = datetime.utcnow() + timedelta(hours=JWT_EXPIRATION_HOURS)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, JWT_SECRET, algorithm=JWT_ALGORITHM)
+    return encoded_jwt
+
+
+def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)) -> int:
+    """JWT 토큰 검증"""
+    try:
+        payload = jwt.decode(credentials.credentials, JWT_SECRET, algorithms=[JWT_ALGORITHM])
+        user_id: int = payload.get("user_id")
+        if user_id is None:
+            raise HTTPException(status_code=401, detail="토큰이 유효하지 않습니다")
+        return user_id
+    except jwt.ExpiredSignatureError:
+        raise HTTPException(status_code=401, detail="토큰이 만료되었습니다")
+    except jwt.JWTError:
+        raise HTTPException(status_code=401, detail="토큰이 유효하지 않습니다")
