+++ eunbi/room-measure/backend-local/outputs/depth_map.npy
