+++ eunbi/room-measure/backend-local/outputs/depth_map_output.png
