+++ eunbi/room-measure/backend-local/outputs/undistorted_image.jpg
