+++ eunbi/room-measure/backend-local/src/detection/corner_intersection_detection.py
