+++ eunbi/room-measure/frontend/README.md
@@ -1,17 +1,17 @@
-# 🏠 Room Measure - 3D 방 측정 및 가구 배치 시뮬레이터
+# Room Measure - 3D 방 측정 및 가구 배치 시뮬레이터
 
 이 프로젝트는 부동산 방 사진을 기반으로 **3D 공간 측정**과 **실시간 가구 배치 시뮬레이션**을 제공하는 고도화된 웹 애플리케이션입니다. 
 
-## 📋 주요 특징
-- 📸 **이미지 기반 방 크기 추정** - OpenCV 왜곡 보정 적용
-- 🎮 **실시간 3D 가구 배치** - Three.js 기반 인터랙티브 환경  
-- 🤖 **AI 인테리어 생성** - 스타일 기반 자동 인테리어 제안
-- 📱 **반응형 디자인** - 모바일/태블릿/데스크톱 지원
-- 🔧 **메인 웹페이지 통합 준비완료** - 모듈화된 컴포넌트 구조
+## 주요 특징
+- **이미지 기반 방 크기 추정** - OpenCV 왜곡 보정 적용
+- **실시간 3D 가구 배치** - Three.js 기반 인터랙티브 환경  
+- **AI 인테리어 생성** - 스타일 기반 자동 인테리어 제안
+- **반응형 디자인** - 모바일/태블릿/데스크톱 지원
+- **메인 웹페이지 통합 준비완료** - 모듈화된 컴포넌트 구조
 
 ---
 
-## ✅ 주요 기능
+## 주요 기능
 
 ### 1. 사진 업로드 및 왜곡 보정
 
@@ -39,7 +39,7 @@
 
 ---
 
-## 🧱 기술 스택
+## 기술 스택
 
 ### Frontend
 - **React 18** - 최신 React with Hooks & Concurrent Features
@@ -61,7 +61,7 @@
 
 ---
 
-## 🚀 실행 방법
+## 실행 방법
 
 ### 1. 백엔드 실행 (FastAPI)
 
@@ -76,7 +76,7 @@
 
 ---
 
-## 📦 주요 디렉토리 구조
+## 주요 디렉토리 구조
 
     room-measure/
     ├── backend/
@@ -93,7 +93,7 @@
 
 ---
 
-## 🎯 향후 확장 가능성
+## 향후 확장 가능성
 
 - RoomNet, HorizonNet 등으로 자동 평면 인식 기능 추가
 - 방 구조 2D 시각화 기능
@@ -101,7 +101,7 @@
 
 ---
 
-## 🔗 메인 웹페이지 통합 가이드
+## 메인 웹페이지 통합 가이드
 
 ### 통합 준비사항
 이 프론트엔드는 메인 웹페이지에 **임베드 가능하도록 설계**되었습니다.
@@ -151,6 +151,6 @@ cp -r dist/* /var/www/main-website/room-measure/
 
 ---
 
-## ✨ 프로젝트 목적
+## 프로젝트 목적
 
 부동산 매물의 광각 사진만으로도 **정확한 방 크기 측정**과 **실시간 가구 배치 시뮬레이션**을 통해 사용자에게 몰입감 있는 공간 경험을 제공합니다. 메인 웹페이지의 핵심 기능으로 통합되어 사용자 참여도와 서비스 가치를 높이는 것이 목표입니다.
